RDCC-3011: Fixed check style errors

+++ src/test/java/uk/gov/hmcts/reform/lrdapi/domain/BuildingLocationTest.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.lrdapi.domain;
 
 import org.junit.Test;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -68,7 +67,7 @@ public class BuildingLocationTest {
 
         BuildingLocation buildingLocation = new BuildingLocation();
 
-        buildingLocation.setBuildingLocationId(Long.valueOf(1));
+        buildingLocation.setBuildingLocationId(1L);
         buildingLocation.setEpimmsId("epimmsId");
         buildingLocation.setBuildingLocationName("buildingLocationName");
         buildingLocation.setBuildingLocationStatus("LIVE");
@@ -83,7 +82,7 @@ public class BuildingLocationTest {
         buildingLocation.setCreated(now);
         buildingLocation.setLastUpdated(now);
 
-        assertThat(buildingLocation.getBuildingLocationId()).isEqualTo(BigInteger.valueOf(1));
+        assertThat(buildingLocation.getBuildingLocationId()).isEqualTo(1L);
         assertThat(buildingLocation.getEpimmsId()).isEqualTo("epimmsId");
         assertThat(buildingLocation.getBuildingLocationName()).isEqualTo("buildingLocationName");
         assertThat(buildingLocation.getBuildingLocationStatus()).isEqualTo("LIVE");
